+++ Domain/Alerts/Alert.cs
@@ -1,0 +1,6 @@
+namespace walle.Domain{
+    public abstract class Alert
+    {
+        public abstract string GetMessage();
+    }
+}
