+++ Domain/Alerts/AllGoodAlert.cs
@@ -1,0 +1,11 @@
+namespace walle.Domain{
+    public class AllGoodAlert : Alert
+    {
+        public AllGoodAlert() { }
+
+        public override string GetMessage()
+        {
+            return "I'm a happy Alert. Success.";
+        }
+    }
+}
