+++ Domain/Analytics/IAnalytic.cs
@@ -1,0 +1,6 @@
+public interface IAnalytic
+{
+    AnalyticType AnalyticType { get; }
+    decimal Get(Currency currency, DateTime? date = null);
+    void Update(Transaction txn);
+}
