+++ Domain/enum/AlertType.cs
@@ -1,0 +1,6 @@
+namespace walle.Domain.Enums{
+    public enum AlertType {
+        BalanceBelow,
+        ValueExceeds
+    }
+}
