+++ Domain/enum/AnalyticType.cs
@@ -1,0 +1,8 @@
+namespace walle.Domain{
+    public enum AnalyticType
+    {
+        NetBalance,
+        DailyAverageTransactionValue,
+        HighestValueTransaction
+    }
+}
