+++ Domain/enum/Currency.cs
@@ -1,0 +1,9 @@
+namespace walle.Domain {
+    public enum Currency
+    {
+        USD,
+        EUR,
+        GBP,
+        IRC
+    }
+}
