+++ Domain/enum/FraudDetectionType.cs
@@ -1,0 +1,9 @@
+namespace walle.Domain.Enums
+{
+    public enum FraudDetectionType
+    {
+        RateLimitExceeded,
+        SuspiciousRoundTrip,
+        AnomalousTransaction
+    }
+}
