+++ Domain/enum/TransactionType.cs
@@ -1,0 +1,8 @@
+namespace walle.Domain.Enums
+{
+    public enum TransactionType
+    {
+        Deposit,
+        Withdrawal
+    }
+}
