+++ ServiceLayer/Interfaces/IRateService.cs
@@ -1,0 +1,6 @@
+namespace walle.ServiceLayer{
+    public interface IRateService
+    {
+        decimal GetRate(Currency baseC, Currency origC);
+    }
+}
