+++ ServiceLayer/Interfaces/IWalletTransactionFraudDetector.cs
@@ -1,0 +1,9 @@
+using walle.Domain;
+
+namespace walle.ServiceLayer
+{
+    public interface IWalletTransactionFraudDetector
+    {
+        (bool, FraudDetectionException?) CheckAndAddEntry(Transaction txn);
+    }
+}
